Remove old RestoreModule classes
