+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/restore/restoremnemonic/RestoreMnemonicModule.kt
@@ -4,20 +4,31 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import io.horizontalsystems.bankwallet.core.App
 import io.horizontalsystems.bankwallet.core.managers.PassphraseValidator
+import io.horizontalsystems.bankwallet.entities.AccountType
 
 object RestoreMnemonicModule {
 
     class Factory : ViewModelProvider.Factory {
         @Suppress("UNCHECKED_CAST")
         override fun <T : ViewModel> create(modelClass: Class<T>): T {
-            val service = RestoreMnemonicService(
-                App.accountFactory,
-                App.wordsManager,
-                PassphraseValidator()
-            )
-
-            return RestoreMnemonicViewModel(service, listOf(service)) as T
+            return RestoreMnemonicViewModel(App.accountFactory, PassphraseValidator(), App.wordsManager, App.thirdKeyboardStorage) as T
         }
     }
+
+    data class UiState(
+        val name: String = "",
+        val defaultName: String,
+        val passphraseEnabled: Boolean = false,
+        val passphrase: String = "",
+        val passphraseError: String? = null,
+        val words: List<WordItem> = listOf(),
+        val invalidWords: List<WordItem> = listOf(),
+        val invalidWordRanges: List<IntRange> = listOf(),
+        val error: String? = null,
+        val accountType: AccountType? = null
+    )
+
+    data class WordItem(val word: String, val range: IntRange)
+    data class State(val allItems: List<WordItem>, val invalidItems: List<WordItem>)
 
 }
