+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/restore/restoremnemonic/RestoreMnemonicViewModel.kt
@@ -1,121 +1,128 @@
 package io.horizontalsystems.bankwallet.modules.restore.restoremnemonic
 
-import androidx.lifecycle.LiveDataReactiveStreams
-import androidx.lifecycle.MutableLiveData
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import io.horizontalsystems.bankwallet.R
-import io.horizontalsystems.bankwallet.core.Clearable
+import io.horizontalsystems.bankwallet.core.IAccountFactory
+import io.horizontalsystems.bankwallet.core.managers.PassphraseValidator
+import io.horizontalsystems.bankwallet.core.managers.WordsManager
 import io.horizontalsystems.bankwallet.core.providers.Translator
 import io.horizontalsystems.bankwallet.entities.AccountType
-import io.horizontalsystems.bankwallet.modules.swap.settings.Caution
-import io.horizontalsystems.core.SingleLiveEvent
-import io.reactivex.BackpressureStrategy
-import io.reactivex.disposables.CompositeDisposable
+import io.horizontalsystems.bankwallet.modules.restore.restoremnemonic.RestoreMnemonicModule.UiState
+import io.horizontalsystems.bankwallet.modules.restore.restoremnemonic.RestoreMnemonicModule.WordItem
+import io.horizontalsystems.core.CoreApp
+import io.horizontalsystems.core.IThirdKeyboard
+import io.horizontalsystems.hdwalletkit.Mnemonic
 
-class RestoreMnemonicViewModel(private val service: RestoreMnemonicService, private val clearables: List<Clearable>) : ViewModel() {
-
-    private val disposables = CompositeDisposable()
-
-    val inputsVisibleLiveData = LiveDataReactiveStreams.fromPublisher(service.passphraseEnabledObservable.toFlowable(BackpressureStrategy.BUFFER))
-    val passphraseCautionLiveData = MutableLiveData<Caution?>()
-    val clearInputsLiveEvent = SingleLiveEvent<Unit>()
-
-    val invalidRangesLiveData = MutableLiveData<List<IntRange>>()
-    val proceedLiveEvent = SingleLiveEvent<Pair<String, AccountType>>()
-    val errorLiveData = MutableLiveData<Throwable>()
-
-    val placeholderLiveData = MutableLiveData(service.defaultName)
+class RestoreMnemonicViewModel(
+    accountFactory: IAccountFactory,
+    private val passphraseValidator: PassphraseValidator,
+    private val wordsManager: WordsManager,
+    private val thirdKeyboardStorage: IThirdKeyboard
+) : ViewModel() {
 
     private val regex = Regex("\\S+")
-    private var state = State(listOf(), listOf())
+    private val defaultName = accountFactory.getNextAccountName()
 
-    override fun onCleared() {
-        clearables.forEach(Clearable::clear)
-        disposables.clear()
+    val isThirdPartyKeyboardAllowed: Boolean
+        get() = CoreApp.thirdKeyboardStorage.isThirdPartyKeyboardAllowed
+
+    val resolvedName: String
+        get() = uiState.name.ifBlank { defaultName }
+
+    var uiState by mutableStateOf(UiState(defaultName = defaultName))
+        private set
+
+    fun onEnablePassphrase(enabled: Boolean) {
+        uiState = uiState.copy(passphraseEnabled = enabled, passphrase = "", passphraseError = null)
+    }
+
+    fun onEnterPassphrase(passphrase: String) {
+        uiState = if (passphraseValidator.validate(passphrase)) {
+            uiState.copy(passphrase = passphrase, passphraseError = null)
+        } else {
+            uiState.copy(
+                passphrase = passphrase,
+                passphraseError = Translator.getString(R.string.CreateWallet_Error_PassphraseForbiddenSymbols)
+            )
+        }
+    }
+
+    fun onEnterMnemonicPhrase(text: String, cursorPosition: Int) {
+        val allItems = wordItems(text)
+        val invalidItems = allItems.filter {
+            !wordsManager.isWordValid(it.word)
+        }
+
+        val invalidWordRanges = invalidItems.filter {
+            val hasCursor = it.range.contains(cursorPosition - 1)
+            !hasCursor || !wordsManager.isWordPartiallyValid(it.word)
+        }.map {
+            it.range
+        }
+
+        uiState = uiState.copy(words = allItems, invalidWords = invalidItems, invalidWordRanges = invalidWordRanges)
+    }
+
+    fun onEnterName(name: String) {
+        uiState = uiState.copy(name = name)
+    }
+
+    fun onProceed() {
+        val passphrase = uiState.passphrase
+        val words = uiState.words.map { it.word }
+        val invalidWords = uiState.invalidWords
+
+        uiState = when {
+            invalidWords.isNotEmpty() -> {
+                uiState.copy(invalidWordRanges = invalidWords.map { it.range })
+            }
+            words.size !in (Mnemonic.EntropyStrength.values().map { it.wordCount }) -> {
+                uiState.copy(
+                    error = Translator.getString(R.string.Restore_Error_MnemonicWordCount, words.size)
+                )
+            }
+            uiState.passphraseError != null -> {
+                uiState
+            }
+            uiState.passphraseEnabled && passphrase.isBlank() -> {
+                uiState.copy(
+                    passphraseError = Translator.getString(R.string.Restore_Error_EmptyPassphrase)
+                )
+            }
+            else -> {
+                try {
+                    wordsManager.validateChecksum(words)
+                    val accountType = AccountType.Mnemonic(words, passphrase)
+
+                    uiState.copy(error = null, accountType = accountType)
+
+                } catch (checksumException: Exception) {
+                    uiState.copy(
+                        error = Translator.getString(R.string.Restore_InvalidChecksum)
+                    )
+                }
+            }
+        }
+    }
+
+    fun onSelectCoinsShown() {
+        uiState = uiState.copy(accountType = null)
+    }
+
+    fun onErrorShown() {
+        uiState = uiState.copy(error = null)
+    }
+
+    fun onAllowThirdPartyKeyboard() {
+        thirdKeyboardStorage.isThirdPartyKeyboardAllowed = true
     }
 
     private fun wordItems(text: String): List<WordItem> {
         return regex.findAll(text.lowercase())
-                .map { WordItem(it.value, it.range) }
-                .toList()
+            .map { WordItem(it.value, it.range) }
+            .toList()
     }
-
-    private fun syncState(text: String) {
-        val allItems = wordItems(text)
-        val invalidItems = allItems.filter {
-            !service.isWordValid(it.word)
-        }
-
-        state = State(allItems, invalidItems)
-    }
-
-    private fun clearInputs() {
-        clearInputsLiveEvent.postValue(Unit)
-        clearCautions()
-
-        service.passphrase = ""
-    }
-
-    private fun clearCautions() {
-        if (passphraseCautionLiveData.value != null) {
-            passphraseCautionLiveData.postValue(null)
-        }
-    }
-
-    fun onNameChange(name: String) {
-        service.name = name
-    }
-
-    fun onTextChange(text: String, cursorPosition: Int) {
-        syncState(text)
-
-        val nonCursorInvalidItems = state.invalidItems.filter {
-            val hasCursor = it.range.contains(cursorPosition - 1)
-
-            !hasCursor || !service.isWordPartiallyValid(it.word)
-        }
-
-        invalidRangesLiveData.postValue(nonCursorInvalidItems.map { it.range })
-    }
-
-    fun onTogglePassphrase(enabled: Boolean) {
-        service.passphraseEnabled = enabled
-        clearInputs()
-    }
-
-    fun onChangePassphrase(v: String) {
-        service.passphrase = v
-        clearCautions()
-    }
-
-    fun validatePassphrase(text: String?): Boolean {
-        val valid = service.validatePassphrase(text)
-        if (!valid) {
-            passphraseCautionLiveData.postValue(Caution(Translator.getString(R.string.CreateWallet_Error_PassphraseForbiddenSymbols), Caution.Type.Error))
-        }
-        return valid
-    }
-
-    fun onProceed() {
-        passphraseCautionLiveData.postValue(null)
-
-        if (state.invalidItems.isNotEmpty()) {
-            invalidRangesLiveData.postValue(state.invalidItems.map { it.range })
-            return
-        }
-
-        try {
-            val accountType = service.accountType(state.allItems.map { it.word })
-
-            proceedLiveEvent.postValue(Pair(service.resolvedName, accountType))
-        } catch (t: RestoreMnemonicService.RestoreError.EmptyPassphrase) {
-            passphraseCautionLiveData.postValue(Caution(Translator.getString(R.string.Restore_Error_EmptyPassphrase), Caution.Type.Error))
-        } catch (error: Throwable) {
-            errorLiveData.postValue(error)
-        }
-    }
-
-    data class WordItem(val word: String, val range: IntRange)
-    data class State(val allItems: List<WordItem>, val invalidItems: List<WordItem>)
-
 }
