--- app/src/main/res/layout/fragment_restore_mnemonic.xml
@@ -1,124 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:background="?android:windowBackground"
-    android:clickable="true">
-
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:paddingEnd="6dp"
-        app:menu="@menu/restore_mnemonic_menu"
-        app:navigationIcon="@drawable/ic_back"
-        app:title="@string/Restore_Enter_Key_Title"
-        app:titleTextAppearance="@style/Title3" />
-
-    <androidx.core.widget.NestedScrollView
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_marginTop="?attr/actionBarSize">
-
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:orientation="vertical">
-
-            <TextView
-                android:id="@+id/accountNameTitle"
-                style="@style/Subhead1"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="16dp"
-                android:layout_marginTop="12dp"
-                android:text="@string/Restore_Name"
-                android:textAllCaps="true"
-                android:textColor="@color/grey"
-                />
-
-            <EditText
-                android:id="@+id/accountNameInput"
-                style="@style/Body"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="16dp"
-                android:layout_marginTop="13dp"
-                android:background="@drawable/input_background_themed"
-                android:gravity="top"
-                android:maxLines="1"
-                android:paddingHorizontal="16dp"
-                android:paddingVertical="12dp"
-                />
-
-            <TextView
-                android:id="@+id/keyInputTitle"
-                style="@style/Subhead1"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="16dp"
-                android:layout_marginTop="44dp"
-                android:text="@string/Restore_Key"
-                android:textAllCaps="true"
-                android:textColor="@color/grey"
-                />
-
-            <EditText
-                android:id="@+id/wordsInput"
-                style="@style/Body"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="16dp"
-                android:layout_marginTop="12dp"
-                android:autofillHints=""
-                android:background="@drawable/input_background_themed"
-                android:gravity="top"
-                android:inputType="textMultiLine|textVisiblePassword"
-                android:maxLines="6"
-                android:minHeight="88dp"
-                android:paddingHorizontal="16dp"
-                android:paddingVertical="12dp" />
-
-            <TextView
-                android:id="@+id/description"
-                style="@style/Subhead2"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginStart="24dp"
-                android:layout_marginTop="6dp"
-                android:layout_marginEnd="24dp"
-                android:text="@string/Restore_Mnemonic_Description"
-                android:textColor="@color/grey" />
-
-            <io.horizontalsystems.views.SettingsViewSwitch
-                android:id="@+id/passphraseToggle"
-                android:layout_width="match_parent"
-                android:layout_height="48dp"
-                android:layout_marginTop="32dp"
-                app:icon="@drawable/ic_key_phrase_20"
-                app:title="@string/Passphrase" />
-
-            <io.horizontalsystems.bankwallet.ui.extensions.InputView
-                android:id="@+id/passphrase"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="16dp"
-                android:layout_marginTop="24dp"
-                android:hint="@string/Passphrase"
-                android:inputType="textPassword" />
-
-            <TextView
-                android:id="@+id/passphraseDescription"
-                style="@style/Subhead2"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:layout_marginHorizontal="16dp"
-                android:layout_marginTop="36dp"
-                android:text="@string/Restore_PassphraseDescription"
-                android:textColor="@color/grey" />
-
-        </LinearLayout>
-    </androidx.core.widget.NestedScrollView>
-
-</FrameLayout>
